fix ts error
