+++ apps/web/lib/actions/partners/invite-partner-from-network.ts
@@ -12,28 +12,9 @@ import ProgramInvite from "@dub/email/templates/program-invite";
 import { prisma } from "@dub/prisma";
 import { waitUntil } from "@vercel/functions";
 import { getProgramOrThrow } from "../../api/programs/get-program-or-throw";
+import { getProgramNetworkInviteEmailDefaults } from "../../network/get-program-network-invite-email-defaults";
 import { authActionClient } from "../safe-action";
 import { throwIfNoPermission } from "../throw-if-no-permission";
-
-export const getProgramNetworkInviteEmailDefaults = ({
-  programName,
-  partnerName: partnerNameProp,
-}: {
-  programName: string;
-  partnerName?: string | null;
-}) => {
-  const trimmedPartnerName = partnerNameProp?.trim();
-  const partnerName =
-    trimmedPartnerName && !trimmedPartnerName.includes("@")
-      ? trimmedPartnerName
-      : "there";
-
-  return {
-    subject: `${programName} invited you to join on Dub Partners`,
-    title: "You're getting noticed!",
-    body: `Hi ${partnerName}, ${programName} found you on the Dub Partner Network and invited you to join their partner program.`,
-  };
-};
 
 export const invitePartnerFromNetworkAction = authActionClient
   .inputSchema(invitePartnerFromNetworkSchema)
